+++ packages/telescope/src/generators/create-rpc-query-clients.ts
@@ -171,6 +171,12 @@ export const plugin = (
 
                           // get all query methods
                           const patterns = item.include?.patterns;
+                          const serviceTypes = item.include?.serviceTypes;
+
+                          if(serviceTypes && !serviceTypes.includes("Query")){
+                            return
+                          }
+
                           const methodKeys = getQueryMethodNames(
                             bundlerFile.package,
                             Object.keys(proto[svcKey].methods ?? {}),