+++ packages/telescope/src/generators/create-types.ts
@@ -64,6 +64,12 @@ export const plugin = (
 
                       // get all query methods
                       const patterns = item.include?.patterns;
+                      const serviceTypes = item.include?.serviceTypes;
+
+                      if(serviceTypes && !serviceTypes.includes("Query")){
+                        return
+                      }
+
                       const methodKeys = getQueryMethodNames(
                         bundlerFile.package,
                         Object.keys(proto[svcKey].methods ?? {}),
@@ -116,8 +122,14 @@ export const plugin = (
                     ...swapKeyValue(item.nameMapping?.Msg ?? {})
                   };
 
-                  // get all query methods
+                  // get all msg methods
                   const patterns = item.include?.patterns;
+                  const serviceTypes = item.include?.serviceTypes;
+
+                  if(serviceTypes && !serviceTypes.includes("Tx")){
+                    return
+                  }
+
                   const methodKeys = getQueryMethodNames(
                     bundlerFile.package,
                     Object.keys(proto.Msg.methods ?? {}),