+++ packages/telescope/src/protod/config.ts
@@ -1,0 +1,30 @@
+import { BufInfo } from "./types";
+
+export const bufInfo: BufInfo[] = [
+  {
+    owner: "protocolbuffers",
+    repo: "wellknowntypes",
+    git: [
+      {
+        owner: "protocolbuffers",
+        repo: "protobuf",
+      },
+    ],
+  },
+  {
+    owner: "cosmos",
+    repo: "gogo-proto",
+    git: [
+      {
+        owner: "cosmos",
+        repo: "gogoproto",
+      },
+    ],
+  },
+];
+
+export const presetProtoDirMapping = {
+  "gogo/protobuf/master": ".",
+  "googleapis/googleapis/master": ".",
+  "protocolbuffers/protobuf/main": "src",
+};